Values should be considered for exclusion from being changed only when the original is different from the default. Closes #47. Thanks for the detailed report manact.
Here's a detailed explanation of what went wrong:

This is kind of complicated to explain, but here goes:

If you want to update or insert a model without first loading it, when setting the data, Jelly does a non-strict (==) comparison between the new value and the current value.

This gets around issue #16 (which, after working on this, I've realized needs to be re-opened). When you set a unique field to the same value it is currently set to, the above behavior ensures that it's not counted as changed and thus validated (erroneously) to be unique.

That's all fine, but the non-strict comparison doesn't work for setting things to 0 and empty strings and such. Those things end up not counting as changes when they should (because we're comparing to the default for the field which defaults to NULL). The fix, then, is to check that the original value is not equal to the default (so we can tell whether or not we're changing).